refactor tests to root utils folder so we can reuse this code for TaikoL1 integration tests
