+++ packages/protocol/test/tokenomics/utils.ts
@@ -1,93 +1,15 @@
 import { BigNumber, ethers } from "ethers";
-import { TaikoL1, TkoToken } from "../../typechain";
 import deployAddressManager from "../utils/addressManager";
-import Proposer from "../utils/proposer";
 import {
     getDefaultL2Signer,
     getL1Provider,
     getL2Provider,
 } from "../utils/provider";
 import createAndSeedWallets from "../utils/seed";
-import sleep from "../utils/sleep";
 import { defaultFeeBase, deployTaikoL1 } from "../utils/taikoL1";
 import { deployTaikoL2 } from "../utils/taikoL2";
 import deployTkoToken from "../utils/tkoToken";
 import { ethers as hardhatEthers } from "hardhat";
-import { BlockProposedEvent } from "../../typechain/LibProposing";
-import { expect } from "chai";
-import verifyBlocks from "../utils/verify";
-import EventEmitter from "events";
-import Prover from "../utils/prover";
-import { BlockMetadata } from "../utils/block_metadata";
-
-type ForkChoice = {
-    provenAt: BigNumber;
-    provers: string[];
-    blockHash: string;
-};
-
-type BlockInfo = {
-    proposedAt: number;
-    provenAt: number;
-    id: number;
-    parentHash: string;
-    blockHash: string;
-    forkChoice: ForkChoice;
-    deposit: BigNumber;
-    proposer: string;
-};
-
-async function sleepUntilBlockIsVerifiable(
-    taikoL1: TaikoL1,
-    id: number,
-    provenAt: number
-) {
-    const delay = await taikoL1.getUncleProofDelay(id);
-    const delayInMs = delay.mul(1000);
-    await sleep(5 * delayInMs.toNumber()); // TODO: use provenAt, calc difference, etc
-}
-
-async function onNewL2Block(
-    l2Provider: ethers.providers.JsonRpcProvider,
-    blockNumber: number,
-    proposer: Proposer,
-    taikoL1: TaikoL1,
-    proposerSigner: any,
-    tkoTokenL1: TkoToken
-): Promise<{
-    proposedEvent: BlockProposedEvent;
-    newProposerTkoBalance: BigNumber;
-    newBlockFee: BigNumber;
-    newProofReward: BigNumber;
-}> {
-    const block = await l2Provider.getBlock(blockNumber);
-    const receipt = await proposer.commitThenProposeBlock(block);
-    const proposedEvent: BlockProposedEvent = (receipt.events as any[]).find(
-        (e) => e.event === "BlockProposed"
-    );
-
-    const { id, meta } = proposedEvent.args;
-
-    const newProofReward = await taikoL1.getProofReward(
-        new Date().getMilliseconds(),
-        meta.timestamp
-    );
-
-    const newProposerTkoBalance = await tkoTokenL1.balanceOf(
-        await proposerSigner.getAddress()
-    );
-
-    const newBlockFee = await taikoL1.getBlockFee();
-
-    console.log("-------------------proposed----------", id);
-
-    return {
-        proposedEvent,
-        newProposerTkoBalance,
-        newBlockFee,
-        newProofReward,
-    };
-}
 
 const sendTinyEtherToZeroAddress = async (signer: any) => {
     await signer
@@ -199,177 +121,4 @@ async function initTokenomicsFixture(mintTkoToProposer: boolean = true) {
     };
 }
 
-async function verifyBlockAndAssert(
-    taikoL1: TaikoL1,
-    tkoTokenL1: TkoToken,
-    block: BlockInfo,
-    lastProofReward: BigNumber,
-    compareHeaders: boolean = false
-): Promise<{ newProofReward: BigNumber }> {
-    await sleepUntilBlockIsVerifiable(taikoL1, block.id, block.provenAt);
-
-    const isVerifiable = await taikoL1.isBlockVerifiable(
-        block.id,
-        block.parentHash
-    );
-
-    expect(isVerifiable).to.be.eq(true);
-
-    // dont verify first blocks parent hash, because we arent "real L2" in these
-    // tests, the parent hash will be wrong.
-    // if (compareHeaders) {
-    //     const latestHash = await taikoL1.getLatestSyncedHeader();
-    //     expect(latestHash).to.be.eq(block.parentHash);
-    // }
-    const prover = block.forkChoice.provers[0];
-
-    const proverTkoBalanceBeforeVerification = await tkoTokenL1.balanceOf(
-        prover
-    );
-
-    const proposerTkoBalanceBeforeVerification = await tkoTokenL1.balanceOf(
-        block.proposer
-    );
-
-    expect(proposerTkoBalanceBeforeVerification.gt(0)).to.be.eq(true);
-    const verifiedEvent = await verifyBlocks(taikoL1, 1);
-    expect(verifiedEvent).to.be.not.undefined;
-
-    expect(verifiedEvent.args.blockHash).to.be.eq(block.blockHash);
-    expect(verifiedEvent.args.id.eq(block.id)).to.be.eq(true);
-
-    const proverTkoBalanceAfterVerification = await tkoTokenL1.balanceOf(
-        prover
-    );
-
-    // prover should have increased in balance as he received the proof reward.
-    expect(
-        proverTkoBalanceAfterVerification.gt(proverTkoBalanceBeforeVerification)
-    ).to.be.eq(true);
-
-    const newProofReward = await taikoL1.getProofReward(
-        block.proposedAt,
-        block.provenAt
-    );
-
-    // last proof reward should be larger than the new proof reward,
-    // since we have stopped proposing, and slots are growing as we verify.
-    if (lastProofReward.gt(0)) {
-        expect(newProofReward).to.be.lt(lastProofReward);
-    }
-
-    // latest synced header should be our just-verified block hash.
-    const latestHash = await taikoL1.getLatestSyncedHeader();
-    expect(latestHash).to.be.eq(block.blockHash);
-
-    // fork choice should be nullified via _cleanUp in LibVerifying
-    const forkChoice = await taikoL1.getForkChoice(block.id, block.parentHash);
-    expect(forkChoice.provenAt).to.be.eq(BigNumber.from(0));
-    expect(forkChoice.provers).to.be.empty;
-    expect(forkChoice.blockHash).to.be.eq(ethers.constants.HashZero);
-
-    // proposer should be minted their refund of their deposit back after
-    // verification, as long as their balance is > 0;
-    return { newProofReward };
-}
-
-function randEle<T>(arr: T[]): T {
-    return arr[Math.floor(Math.random() * arr.length)];
-}
-
-function newProposerListener(
-    genesisHeight: number,
-    eventEmitter: EventEmitter,
-    l2Provider: ethers.providers.JsonRpcProvider,
-    proposer: Proposer,
-    taikoL1: TaikoL1,
-    tkoTokenL1: TkoToken
-) {
-    return async (blockNumber: number) => {
-        try {
-            if (blockNumber <= genesisHeight) return;
-
-            const { proposedEvent } = await onNewL2Block(
-                l2Provider,
-                blockNumber,
-                proposer,
-                taikoL1,
-                proposer.getSigner(),
-                tkoTokenL1
-            );
-            expect(proposedEvent).not.to.be.undefined;
-
-            eventEmitter.emit(BLOCK_PROPOSED_EVENT, proposedEvent, blockNumber);
-        } catch (e) {
-            eventEmitter.emit("error", e);
-        }
-    };
-}
-
-function newProverListener(
-    prover: Prover,
-    taikoL1: TaikoL1,
-    eventEmitter: EventEmitter
-) {
-    return async (proposedEvent: BlockProposedEvent, blockNumber: number) => {
-        try {
-            const { args } = await prover.prove(
-                await prover.getSigner().getAddress(),
-                proposedEvent.args.id.toNumber(),
-                blockNumber,
-                proposedEvent.args.meta as any as BlockMetadata
-            );
-            const { blockHash, id: blockId, parentHash, provenAt } = args;
-
-            const proposedBlock = await taikoL1.getProposedBlock(
-                proposedEvent.args.id.toNumber()
-            );
-
-            const forkChoice = await taikoL1.getForkChoice(
-                blockId.toNumber(),
-                parentHash
-            );
-
-            expect(forkChoice.blockHash).to.be.eq(blockHash);
-
-            expect(forkChoice.provers[0]).to.be.eq(
-                await prover.getSigner().getAddress()
-            );
-
-            const provedBlock = {
-                proposedAt: proposedBlock.proposedAt.toNumber(),
-                provenAt: provenAt.toNumber(),
-                id: proposedEvent.args.id.toNumber(),
-                parentHash: parentHash,
-                blockHash: blockHash,
-                forkChoice: forkChoice,
-                deposit: proposedBlock.deposit,
-                proposer: proposedBlock.proposer,
-            };
-
-            eventEmitter.emit(BLOCK_PROVEN_EVENT, provedBlock);
-        } catch (e) {
-            eventEmitter.emit("error", e);
-        }
-    };
-}
-
-const BLOCK_PROPOSED_EVENT = "blockProposed";
-const BLOCK_PROVEN_EVENT = "blockProved";
-const BLOCK_VERIFIED_EVENT = "blockVerified";
-
-export {
-    sendTinyEtherToZeroAddress,
-    onNewL2Block,
-    sleepUntilBlockIsVerifiable,
-    initTokenomicsFixture,
-    verifyBlockAndAssert,
-    randEle,
-    newProposerListener,
-    newProverListener,
-    BLOCK_PROPOSED_EVENT,
-    BLOCK_PROVEN_EVENT,
-    BLOCK_VERIFIED_EVENT,
-};
-
-export type { BlockInfo, ForkChoice };
+export { initTokenomicsFixture };
