+++ packages/protocol/test/utils/array.ts
@@ -1,0 +1,5 @@
+function randEle<T>(arr: T[]): T {
+    return arr[Math.floor(Math.random() * arr.length)];
+}
+
+export { randEle };
