+++ packages/protocol/test/utils/propose.ts
@@ -1,8 +1,13 @@
+import { expect } from "chai";
 import { BigNumber, ethers } from "ethers";
 import RLP from "rlp";
-import { TaikoL1 } from "../../typechain";
+import { EventEmitter } from "stream";
+import { TaikoL1, TkoToken } from "../../typechain";
 import { BlockMetadata } from "./block_metadata";
 import { encodeBlockMetadata } from "./encoding";
+import { BLOCK_PROPOSED_EVENT } from "./event";
+import { onNewL2Block } from "./onNewL2Block";
+import Proposer from "./proposer";
 
 const buildProposeBlockInputs = (
     block: ethers.providers.Block,
@@ -44,4 +49,33 @@ const proposeBlock = async (
     return receipt;
 };
 
-export { buildProposeBlockInputs, proposeBlock };
+function newProposerListener(
+    genesisHeight: number,
+    eventEmitter: EventEmitter,
+    l2Provider: ethers.providers.JsonRpcProvider,
+    proposer: Proposer,
+    taikoL1: TaikoL1,
+    tkoTokenL1: TkoToken
+) {
+    return async (blockNumber: number) => {
+        try {
+            if (blockNumber <= genesisHeight) return;
+
+            const { proposedEvent } = await onNewL2Block(
+                l2Provider,
+                blockNumber,
+                proposer,
+                taikoL1,
+                proposer.getSigner(),
+                tkoTokenL1
+            );
+            expect(proposedEvent).not.to.be.undefined;
+
+            eventEmitter.emit(BLOCK_PROPOSED_EVENT, proposedEvent, blockNumber);
+        } catch (e) {
+            eventEmitter.emit("error", e);
+        }
+    };
+}
+
+export { buildProposeBlockInputs, proposeBlock, newProposerListener };
